fix: moving functions
